Major text fixes/bug fixes
New:
Added the ability to not buy any of a selected shop item by typing 0 for the quantity you want to purchase.
Added a cap to how many movement commands you can do at once, you can now only queue up 3 movement commands at a time

Changes:
Changed the viewing info about Oochamon pop up to last 20 seconds rather than 10 seconds.
Changed the tile jump cap to 4
Changed fast battle speed from 0.5 seconds to 1.25 seconds, and normal battle speed from 1.25 seconds to 2.5 seconds.
Made discord movement buttons the default movement method for new players.
Made the controls message by default off, as its not necessary with the discord movement buttons being on by default.
Changed NPCs to use the new npc_id system for their auto generated flags.
Changed stat decreasing/increasing moves to not show the stat amount changed (you can still see it in the battle info button however)
Changed super effective/not very effective to be in line with the amount of damage done (to make it more readable)

Fixes:
Fixed bug where non-attacking moves showed as super effective
Fixed bug where prism item descriptions got cut off after 25 characters, when they should get cut off after 100
Fixed bug where "rest of your team earned X xp" message would show up if you only had 1 oochamon.
Fixed a hard crash with the bot when trying to reload the game if the game thread is deleted manually by a user
Fixed a bug where items would display healing the wrong number.
Fixed numerous issues with vampire healing.

+++ ooch_bot/index.js
@@ -95,17 +95,19 @@ client.on('ready', async () => {
         } else if ((user_profile.player_state !== PlayerState.NotPlaying && user_profile.player_state !== PlayerState.Playspace) ||
             (user_profile.player_state === PlayerState.Playspace && user_profile.settings.discord_move_buttons === true)) {
 
-            await userThread.bulkDelete(100);
             db.profile.set(user, [], 'npc_event_data');
             db.profile.set(user, 0, 'npc_event_pos')
+            if (userThread !== undefined) {
+                await userThread.bulkDelete(100)
 
-            // Setup playspace
-            let playspace_str = await setup_playspace_str(user);
-            await db.profile.set(user, PlayerState.Playspace, 'player_state');
+                // Setup playspace
+                let playspace_str = await setup_playspace_str(user);
+                await db.profile.set(user, PlayerState.Playspace, 'player_state');
 
-            await userThread.send({ content: playspace_str[0], components: playspace_str[1] }).then(msg => {
-                db.profile.set(user, msg.id, 'display_msg_id');
-            });
+                await userThread.send({ content: playspace_str[0], components: playspace_str[1] }).then(msg => {
+                    db.profile.set(user, msg.id, 'display_msg_id');
+                });
+            }
 
             // let warningMsg = await userThread.send({ content: '## The bot has crashed, and your game has soft rebooted to avoid corruption and button issues. No progress has been lost.' });
 
@@ -295,20 +297,24 @@ client.on('messageCreate', async message => {
                     // Do movement stuff
                     let moveMsg = message.content.toLowerCase().replace(/\s+/g, ''); // Remove spaces
                     let splitMoves = moveMsg.split(','); // Split by comma first
+                    let moveCount = 0;
                 
                     let currentDirection = null;
                     let totalDistance = 0;
                 
                     for (let msg of splitMoves) {
+                        if (moveCount >= 3) break; // Stop processing after 3 moves
                         let moveSequence = msg.match(/([wasd])(\d{1,2})?/g); // Match 'w', 'a', 's', 'd' followed optionally by a number
                 
                         if (!moveSequence) continue; // Skip invalid input
                 
                         for (let moveSeq of moveSequence) {
+                            if (moveCount >= 3) break; // Stop further moves if limit reached
                             let direction = moveSeq[0]; // First character (direction)
                             let dist = moveSeq.length > 1 ? parseInt(moveSeq.slice(1)) : 1; // Get distance if provided, otherwise default to 1
                 
                             if (isNaN(dist)) dist = 1; // Safety check for NaN values
+                            if (dist > 4) dist = 4;
                 
                             if (currentDirection === null) {
                                 currentDirection = direction;
@@ -318,6 +324,9 @@ client.on('messageCreate', async message => {
                             } else {
                                 // Move in the previous direction with the total distance
                                 await move(message.channel, message.author.id, currentDirection, totalDistance);
+                                moveCount++;
+
+                                if (moveCount >= 3) break; // Stop further moves if limit reached                    
                 
                                 // Start a new aggregation for the current direction
                                 currentDirection = direction;
@@ -327,7 +336,7 @@ client.on('messageCreate', async message => {
                     }
                 
                     // Call move for the last accumulated direction
-                    if (currentDirection !== null) {
+                    if (currentDirection !== null && moveCount < 3) {
                         await move(message.channel, message.author.id, currentDirection, totalDistance);
                     }
